* fixed gcc for the package name, converted version number checks 	  to use the real version number - not the package name and adapted 	  the stack smash protector extraction to the recent release format
git-svn-id: https://svn.exactcode.de/t2/trunk@3876 c5f82cb5-29bc-0310-9cd0-bff59a50e3bc

+++ package/base/gcc/apply-protector.sh
@@ -5,15 +5,19 @@ if [ "$pfile" ] ; then
 	tar --use-compress-program=bzip2 \
 	  -xf $archdir/${pfile/.gz/.bz2}
 
-	# Patch protector.dif a bit to apply against current gcc-3
-	#[ $2 = gcc3 ] && patch -p1 < $1/package/base/$2/protector-hotfix.diff
-
 	# be careful if you enable this, you have to respect $pkg.
 	# Set -fstack-protector as default?
 	# [ $ROCKCFG_PKG_GCC[23]_STACKPRO = 1 ] && patch -p0 < protectonly.dif
 
-	patch -p0 < protector.dif
-	mv protector.{c,h} gcc/
+	if [ -f protector.dif ] ; then
+		patch -p0 < protector.dif
+		mv protector.{c,h} gcc/
+	elif [ -f gcc_*.dif ] ; then
+		patch -p0 < gcc_*.dif
+		# the protector.* files already extract into gcc/ ...
+	else
+		abord "Protector patch not found"
+	fi
 else
 	echo "No stack-protector available for $2 ..."
 fi
